trip-log-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pyjwt>=2.8.0
python-dotenv>=1.0.0
